Moved response selection utils to logic adapter class.
